Move date selector to new DateSelector component
