+++ frontend/src/components/SeasonView.tsx
@@ -1,4 +1,4 @@
-import { useEffect, useMemo, useState } from "react";
+import { useMemo, useState } from "react";
 import type { GrandPrix } from "@/types";
 import { useQuery } from "@tanstack/react-query";
 import ConstructorPositionsTable from "./ConstructorPositionsTable";
@@ -8,19 +8,20 @@ import {
   generateDriverLineChartData,
   generateConstructorLineChartData,
 } from "../utils/generateDataSets";
-import { buildRaceDateString } from "../utils/stringUtils";
 import DriverSeasonLineGraph from "./DriverSeasonLineGraph";
 import ConstructorLineGraph from "./ConstructorLineGraph";
 import ChampionshipToggleSwitch from "./ChampionshipToggleSwitch";
 import { API_BASE_URL } from "../constants/urls";
 import SeasonSelector from "./SeasonSelector";
 import ClearDatesButton from "./ClearDatesButton";
+import DateSelector from "./DateSelector";
 
 const SeasonView = () => {
   const [isTeam, setIsTeam] = useState(false);
   const [year, setYear] = useState("2025");
   const [startDate, setStartDate] = useState("");
   const [endDate, setEndDate] = useState("");
+  
   const { isPending, error, data } = useQuery({
     queryKey: ["season", year],
     queryFn: () =>
@@ -32,21 +33,6 @@ const SeasonView = () => {
       ),
   });
   const races = (data as GrandPrix[]) || [];
-
-  useEffect(() => {
-    if (races.length > 0) {
-      setStartDate(races[0].date);
-      setEndDate(races[races.length - 1].date);
-    }
-  }, [races]);
-
-  // Filter end date options to only show dates after the startDate.
-  const availableEndDates = useMemo(() => {
-    if (!startDate) return races;
-
-    const startIndex = races.findIndex((race) => race.date === startDate);
-    return startIndex >= 0 ? races.slice(startIndex) : races;
-  }, [races, startDate]);
 
   const processedData = useMemo(
     () => calculateStandings(races, startDate, endDate),
@@ -73,38 +59,13 @@ const SeasonView = () => {
       <div className="px-5 py-1.5 flex flex-col md:flex-row items-start md:items-center space-x-1 md:space-x-2 gap-3 text-md border border-black rounded-2xl bg-gray-100">
         <ChampionshipToggleSwitch isTeam={isTeam} setIsTeam={setIsTeam} />
         <SeasonSelector year={year} setYear={setYear} />
-        <div className="flex gap-1 justify-center items-center">
-          <span className={`text-md`}>Start Date:</span>
-          <select
-            className="border border-gray-400 p-1 rounded-sm text-sm hover:bg-gray-300 cursor-pointer"
-            value={startDate}
-            onChange={(e) => setStartDate(e.target.value)}
-          >
-            {races
-              .filter((race) => !race.is_sprint)
-              .map((race, idx) => (
-                <option value={race.date} key={`s${race.round}${idx}`}>
-                  {buildRaceDateString(race)}
-                </option>
-              ))}
-          </select>
-        </div>
-        <div className="flex gap-1 justify-center items-center">
-          <span className={`text-md`}>End Date:</span>
-          <select
-            className="border border-gray-400 p-1 rounded-sm text-sm hover:bg-gray-300 cursor-pointer"
-            value={endDate}
-            onChange={(e) => setEndDate(e.target.value)}
-          >
-            {availableEndDates
-              .filter((race) => !race.is_sprint)
-              .map((race, idx) => (
-                <option value={race.date} key={`e${race.round}${idx}`}>
-                  {buildRaceDateString(race)}
-                </option>
-              ))}
-          </select>
-        </div>
+        <DateSelector
+          races={races}
+          startDate={startDate}
+          endDate={endDate}
+          setStartDate={setStartDate}
+          setEndDate={setEndDate}
+        />
         <ClearDatesButton
           races={races}
           setStartDate={setStartDate}
